refactor: 주소 관련 로직 모듈화
