+++ src/features/user/address/hooks/useAddressForm.ts
@@ -1,7 +1,8 @@
-import { createUserAddress } from '@/features/user/address/apis';
+import {
+  useSearchAddress,
+  useUserAddressMutation,
+} from '@/features/user/address';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { useEffect } from 'react';
 import { useForm } from 'react-hook-form';
 import { z } from 'zod';
 
@@ -27,17 +28,8 @@ export default function useAddressForm(onClose: () => void) {
     formState: { errors },
     setValue,
   } = useForm<AddressForm>({ resolver: zodResolver(schema) });
-
-  //FIXME: 분리
-  const queryClient = useQueryClient();
-  const { mutate: mutateUserAddress } = useMutation({
-    mutationFn: createUserAddress,
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['addresses'] });
-      queryClient.invalidateQueries({ queryKey: ['defaultAddress'] });
-      onClose();
-    },
-  });
+  const { mutate: mutateUserAddress } = useUserAddressMutation();
+  const { handlePostcode } = useSearchAddress(setValue);
 
   const onSubmit = (data: AddressForm) => {
     const formattedData = {
@@ -46,52 +38,7 @@ export default function useAddressForm(onClose: () => void) {
     };
 
     mutateUserAddress(formattedData);
-  };
-
-  useEffect(() => {
-    const script = document.createElement('script');
-    script.src =
-      '//t1.daumcdn.net/mapjsapi/bundle/postcode/prod/postcode.v2.js';
-    script.async = true;
-    document.head.appendChild(script);
-
-    return () => {
-      document.head.removeChild(script);
-    };
-  }, []);
-
-  const handlePostcode = (e: React.MouseEvent) => {
-    e.preventDefault();
-
-    new window.daum.Postcode({
-      oncomplete: (data: {
-        zonecode: string;
-        roadAddress: string;
-        bname: string;
-        buildingName: string;
-        apartment: string;
-        autoRoadAddress: string;
-        autoJibunAddress: string;
-      }) => {
-        const roadAddr = data.roadAddress;
-        let extraRoadAddr = '';
-
-        if (data.bname !== '' && /[동|로|가]$/g.test(data.bname)) {
-          extraRoadAddr += data.bname;
-        }
-        if (data.buildingName !== '' && data.apartment === 'Y') {
-          extraRoadAddr +=
-            extraRoadAddr !== '' ? `, ${data.buildingName}` : data.buildingName;
-        }
-        if (extraRoadAddr !== '') {
-          extraRoadAddr = ` (${extraRoadAddr})`;
-        }
-
-        setValue('zonecode', data.zonecode);
-        setValue('roadAddress', roadAddr);
-        setValue('extraAddress', extraRoadAddr);
-      },
-    }).open();
+    onClose();
   };
 
   return { handlePostcode, register, handleSubmit, onSubmit, errors };
