+++ src/features/user/address/index.ts
@@ -5,6 +5,8 @@ export { default as AddressList } from './components/AddressList';
 export { default as useAddressList } from './hooks/useAddressList';
 export { default as useAddressForm } from './hooks/useAddressForm';
 export { default as useDeliverySummary } from './hooks/useDeliverySummary';
+export { default as useUserAddressMutation } from './hooks/useUserAddressMutation';
+export { default as useSearchAddress } from './hooks/useSearchAddress';
 
 export * from './types';
 export * from './apis';
